requirements.txt
pysml>=0.0.2
pyserial
humanfriendly
requests
PyYAML
smbus2
construct
curses-menu